requirements.txt
# Core Dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
websockets>=10.1
torch>=1.9.0
numpy>=1.21.0
pydantic>=1.8.2
pyyaml>=5.4.1
python-multipart>=0.0.5
python-dotenv>=0.19.0
aiohttp>=3.7.4

# Optional/Development
pytest>=6.2.5
jupyterlab>=3.2.0
ipython>=7.29.0
matplotlib>=3.4.3
ipywidgets>=7.6.5
websocket-client>=1.8.0
